replaced ThreadedSE with HashedSE
