+++ src/main/java/org/threadly/litesockets/ThreadedSocketExecuter.java
@@ -1,47 +1,29 @@
 package org.threadly.litesockets;
 
+import java.io.IOException;
 import java.nio.channels.CancelledKeyException;
+import java.nio.channels.ClosedSelectorException;
 import java.nio.channels.DatagramChannel;
 import java.nio.channels.SelectionKey;
-import java.nio.channels.ServerSocketChannel;
-import java.util.Arrays;
+import java.nio.channels.Selector;
+import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.ScheduledExecutorService;
 
 import org.threadly.concurrent.ConfigurableThreadFactory;
 import org.threadly.concurrent.SingleThreadScheduler;
 import org.threadly.concurrent.SubmitterExecutor;
 import org.threadly.concurrent.SubmitterScheduler;
-import org.threadly.concurrent.future.FutureUtils;
-import org.threadly.concurrent.future.ListenableFuture;
 import org.threadly.concurrent.wrapper.KeyDistributedExecutor;
 import org.threadly.concurrent.wrapper.compatibility.ScheduledExecutorServiceWrapper;
 import org.threadly.litesockets.utils.IOUtils;
 import org.threadly.util.ArgumentVerifier;
 import org.threadly.util.ExceptionUtils;
 
-/**
- * <p>This is a mutliThreaded implementation of a {@link SocketExecuter}.  It uses separate threads to perform Accepts, Reads and Writes.  
- * Constructing this will create 3 additional threads.  Generally only one of these will ever be needed in a process.</p>
- * 
- * <p>This is generally the {@link SocketExecuter} implementation you want to use for servers, especially if they have to deal with more
- * then just a few connections.  See {@link NoThreadSocketExecuter} for a more efficient implementation when not dealing with many connections.</p>
- * 
- * @author lwahlmeier
- *
- */
 public class ThreadedSocketExecuter extends SocketExecuterCommonBase {
+  private final SelectorThread[] clientSelectors;
   private final KeyDistributedExecutor clientDistributer;
-  private final SingleThreadScheduler localAcceptScheduler;
-  private final SingleThreadScheduler localReadScheduler;
-  private final SingleThreadScheduler localWriteScheduler;
-
-  private volatile long readThreadID;
-
-  private final AcceptRunner acceptor = new AcceptRunner();
-  private final ReadRunner reader = new ReadRunner();
-  private final WriteRunner writer = new WriteRunner();
-
-
+  private final int selectors;
+  
   /**
    * <p>This constructor creates its own {@link SingleThreadScheduler} Threadpool to use for client operations.  This is generally 
    * not recommended unless you are not doing many socket connections/operations.  You should really use your own multiThreaded 
@@ -52,7 +34,7 @@ public class ThreadedSocketExecuter extends SocketExecuterCommonBase {
         new ConfigurableThreadFactory(
             "SocketClientThread", false, true, Thread.currentThread().getPriority(), null, null)));
   }
-
+  
   /**
    * <p>This is provided to allow people to use java's generic threadpool scheduler {@link ScheduledExecutorService}.</p>
    * 
@@ -61,38 +43,103 @@ public class ThreadedSocketExecuter extends SocketExecuterCommonBase {
   public ThreadedSocketExecuter(final ScheduledExecutorService exec) {
     this(new ScheduledExecutorServiceWrapper(exec));
   }
-
+  
   /**
-   * <p>Here you can provide a {@link ScheduledExecutorService} for this {@link SocketExecuter}.  This will be used
+   * <p>Here you can provide a {@link SubmitterScheduler} for this {@link SocketExecuter}.  This will be used
    * on accept, read, and close callback events.</p>
    * 
-   * @param exec the {@link ScheduledExecutorService} to be used for client/server callbacks.
+   * @param scheduler the {@link SubmitterScheduler} to be used for client/server callbacks.
    */
-  public ThreadedSocketExecuter(final SubmitterScheduler exec) {
-    this(exec, Integer.MAX_VALUE);
+  public ThreadedSocketExecuter(final SubmitterScheduler scheduler) {
+    this(scheduler, Integer.MAX_VALUE, -1);
   }
   
-  public ThreadedSocketExecuter(final SubmitterScheduler exec, int maxTasksPerCycle) {
-    super(new SingleThreadScheduler(new ConfigurableThreadFactory("SocketAccept", false, true, Thread.currentThread().getPriority(), null, null)),
-        new SingleThreadScheduler(new ConfigurableThreadFactory("SocketReader", false, true, Thread.currentThread().getPriority(), null, null)),
-        new SingleThreadScheduler(new ConfigurableThreadFactory("SocketWriter", false, true, Thread.currentThread().getPriority(), null, null)),
-        exec);
-    localAcceptScheduler = (SingleThreadScheduler) this.acceptScheduler;
-    localReadScheduler = (SingleThreadScheduler) this.readScheduler;
-    localWriteScheduler = (SingleThreadScheduler) this.writeScheduler;
-    clientDistributer = new KeyDistributedExecutor(schedulerPool);
+  /**
+   * <p>Creates a ThreadedSocketExecuter with different max task cycle</p> 
+   * 
+   * @param scheduler the {@link SubmitterScheduler} to be used for client/server callbacks.
+   * @param maxTasksPerCycle the max number of tasks to run on a clients thread before returning the thread back to the pool.
+   * @param numberOfSelectors the number of selector threads to run.  Default is core/2.
+   */
+  public ThreadedSocketExecuter(final SubmitterScheduler scheduler, final int maxTasksPerCycle) {
+    this(scheduler, maxTasksPerCycle, -1);
+  }
+  
+  /**
+   * <p>Creates a ThreadedSocketExecuter with different max task cycle and Selector Thread defaults.</p> 
+   * 
+   * @param scheduler the {@link SubmitterScheduler} to be used for client/server callbacks.
+   * @param maxTasksPerCycle the max number of tasks to run on a clients thread before returning the thread back to the pool.
+   * @param numberOfSelectors the number of selector threads to run.  Default is core/2.
+   */
+  public ThreadedSocketExecuter(final SubmitterScheduler scheduler, final int maxTasksPerCycle, final int numberOfSelectors) {
+    super(scheduler);
+    int ps = -1;
+    if(numberOfSelectors == -1) {
+       ps = Runtime.getRuntime().availableProcessors()/2;
+    } else {
+      ps = numberOfSelectors;
+    }
+    clientSelectors = new SelectorThread[ps];
+    clientDistributer = new KeyDistributedExecutor(schedulerPool, maxTasksPerCycle);
+    this.selectors = ps;
+  }
+  
+  private SelectorThread getSelectorFor(Object obj) {
+    if(selectors == 1) {
+      return clientSelectors[0];
+    } 
+    return clientSelectors[obj.hashCode()%selectors];
+  }
+
+  @Override
+  public SubmitterExecutor getExecutorFor(Object obj) {
+    return clientDistributer.getExecutorForKey(obj);
+  }
+
+  @Override
+  public void setClientOperations(Client client) {
+    ArgumentVerifier.assertNotNull(client, "Client");
+    if(!clients.containsKey(client.getChannel())) {
+      return;
+    }
+    final SelectorThread st = getSelectorFor(client);
+    st.addClient(client);
+  }
+  
+  @Override
+  public void startListening(final Server server) {
+    if(checkServer(server)) {
+      final SelectorThread st = getSelectorFor(server);
+      st.addServer(server);
+    }
+  }
+
+  @Override
+  public void stopListening(final Server server) {
+    if(checkServer(server)) {
+      final SelectorThread st = getSelectorFor(server);
+      st.removeServer(server);
+    }
+  }
+
+  @Override
+  public void setUDPServerOperations(UDPServer udpServer, boolean enable) {
+    if(checkServer(udpServer)) {
+      final SelectorThread st = getSelectorFor(udpServer);
+      if(enable) {
+        st.addServer(udpServer);
+      } else {
+        st.removeServer(udpServer);
+      }
+    }
   }
 
   @Override
   protected void startupService() {
-    super.startIfNotStarted();
-
-    acceptSelector = openSelector();
-    readSelector = openSelector();
-    writeSelector = openSelector();
-    acceptScheduler.execute(acceptor);
-    readScheduler.execute(reader);
-    writeScheduler.execute(writer);
+    for(int i=0; i<selectors; i++) {
+      clientSelectors[i] = new SelectorThread(i);
+    }
   }
 
   @Override
@@ -103,194 +150,187 @@ public class ThreadedSocketExecuter extends SocketExecuterCommonBase {
     for(final Server server: servers.values()) {
       IOUtils.closeQuietly(server);
     }
-    closeSelector(localAcceptScheduler, acceptSelector);
-    closeSelector(localReadScheduler, readSelector);
-    closeSelector(localWriteScheduler, writeSelector);
-  }  
-
-  @Override
-  public void setClientOperations(final Client client) {
-    ArgumentVerifier.assertNotNull(client, "Client");
-    if(!clients.containsKey(client.getChannel())) {
-      clients.remove(client.getChannel());
-      return;
-    }
-
-    synchronized(client) {
-      if(client.isClosed()) {
-        clients.remove(client.getChannel());
-        ListenableFuture<?> lf = readScheduler.submit(new RemoveFromSelector(readSelector, client));
-        ListenableFuture<?> lf2 = writeScheduler.submit(new RemoveFromSelector(writeSelector, client));
-        FutureUtils.makeCompleteFuture(Arrays.asList(lf, lf2)).addListener(new Runnable() {
-          @Override
-          public void run() {
-            IOUtils.closeQuietly(client.getChannel());
-          }});
-      } else if(!client.getChannel().isConnected() && client.getChannel().isConnectionPending()) {
-        readScheduler.execute(new AddToSelector(readScheduler, client, readSelector, SelectionKey.OP_CONNECT));
-        writeScheduler.execute(new AddToSelector(writeScheduler, client, writeSelector, 0));
-      } else if(client.canWrite() && client.canRead()) {
-        writeScheduler.execute(new AddToSelector(writeScheduler, client, writeSelector, SelectionKey.OP_WRITE));
-        readScheduler.execute(new AddToSelector(readScheduler, client, readSelector, SelectionKey.OP_READ));
-      } else if (client.canRead()){
-        readScheduler.execute(new AddToSelector(readScheduler, client, readSelector, SelectionKey.OP_READ));
-        writeScheduler.execute(new AddToSelector(writeScheduler, client, writeSelector, 0));
-      } else if (client.canWrite()){
-        writeScheduler.execute(new AddToSelector(writeScheduler, client, writeSelector, SelectionKey.OP_WRITE));
-        readScheduler.execute(new AddToSelector(readScheduler, client, readSelector, 0));
-      } else {
-        writeScheduler.execute(new AddToSelector(writeScheduler, client, writeSelector, 0));
-        readScheduler.execute(new AddToSelector(readScheduler, client, readSelector, 0));
-      }
-    }
-    readSelector.wakeup();
-    writeSelector.wakeup();
-  }
-
-  @Override
-  public void setUDPServerOperations(final UDPServer udpServer, final boolean enable) {
-    if(checkServer(udpServer)) {
-      if(enable) {
-        readScheduler.execute(new AddToSelector(readScheduler, udpServer, readSelector, SelectionKey.OP_READ));
-        readSelector.wakeup();
-        if(udpServer.needsWrite()) {
-          writeScheduler.execute(new AddToSelector(writeScheduler, udpServer, writeSelector, SelectionKey.OP_WRITE));
-          writeSelector.wakeup();
-        } else {
-          writeScheduler.execute(new AddToSelector(writeScheduler, udpServer, writeSelector, 0));
-          writeSelector.wakeup();
-        }
-      } else {
-        readScheduler.execute(new AddToSelector(readScheduler, udpServer, readSelector, 0));
-        writeScheduler.execute(new AddToSelector(writeScheduler, udpServer, writeSelector, 0));
-        readSelector.wakeup();
-        writeSelector.wakeup();
-      }
+    for(SelectorThread st: clientSelectors) {
+      st.selector.wakeup();
+      st.selector.wakeup();
+      
+      IOUtils.closeQuietly(st.selector);
     }
   }
-
-  /**
-   * Runnable for the Acceptor thread.  This runs the acceptSelector on the AcceptorThread. 
-   */
-  private class AcceptRunner implements Runnable {
-    @Override
-    public void run() {
-      if(isRunning()) {
+  
+  private class SelectorThread {
+    private final Selector selector;
+    private final Thread thread;
+    private final ConcurrentLinkedQueue<Client> clientsToCheck = new ConcurrentLinkedQueue<>();
+    private final ConcurrentLinkedQueue<Server> serversToAdd = new ConcurrentLinkedQueue<>();
+    private final ConcurrentLinkedQueue<Server> serversToRemove = new ConcurrentLinkedQueue<>();
+    private volatile boolean isAwake = true;
+    
+    public SelectorThread(int id) {
+      selector = openSelector();
+      thread = new Thread(()->doSelect(), "HashedSelector");
+      thread.setDaemon(true);
+      thread.start();
+    }
+    
+    public void addClient(Client client) {
+      clientsToCheck.add(client);
+      if(!isAwake) {
+        isAwake = true;
+        selector.wakeup();
+      }
+    }
+    
+    public void addServer(Server server) {
+      serversToAdd.add(server);
+      if(!isAwake) {
+        isAwake = true;
+        selector.wakeup();
+      }
+    }
+    
+    public void removeServer(Server server) {
+      serversToRemove.add(server);
+      if(!isAwake) {
+        isAwake = true;
+        selector.wakeup();
+      }
+    }
+    
+    @SuppressWarnings("resource")
+    private void processServers() {
+      Server server = serversToAdd.poll();
+      while(server != null) {
         try {
-          acceptSelector.selectedKeys().clear();
-          acceptSelector.select();
-        } catch (Exception e) {
-          ExceptionUtils.handleException(e);
-          stopIfRunning();
-          return;
-        } 
-        if(isRunning()) {
-          for(final SelectionKey sk: acceptSelector.selectedKeys()) {
-            if(sk.isAcceptable()) {
-              final ServerSocketChannel server = (ServerSocketChannel) sk.channel();
-              doServerAccept(servers.get(server));
-            }
-          }
-        }
-        if(isRunning()) {
-          acceptScheduler.execute(this);
-        }
-      }
-    }
-  }
-
-  /**
-   * Runnable for the Read thread.  This runs the readSelector on the ReadThread. 
-   */
-  private class ReadRunner implements Runnable {
-    @Override
-    public void run() {
-      if(isRunning()) {
-        if(readThreadID == 0) {
-          readThreadID = Thread.currentThread().getId();
-        }
-        try {
-          readSelector.selectedKeys().clear();
-          readSelector.select();
-        } catch (Exception e) {
-          ExceptionUtils.handleException(e);
-          stopIfRunning();
-          return;
-        }
-        if(isRunning() && ! readSelector.selectedKeys().isEmpty()) {
-          for(final SelectionKey sk: readSelector.selectedKeys()) {
-            final Client client = clients.get(sk.channel());
-            if(client != null) {
-              try {
-                if(sk.isConnectable()) {
-                  doClientConnect(client, readSelector);
-                  sk.cancel();
-                } else {
-                  doClientRead(client, readSelector);
-                }
-              } catch(CancelledKeyException e) {
-                IOUtils.closeQuietly(client);
-                ExceptionUtils.handleException(e);
-              }
+          if(server.getServerType() == WireProtocol.TCP) {
+            server.getSelectableChannel().register(selector, SelectionKey.OP_ACCEPT);
+          } else if(server.getServerType() == WireProtocol.UDP) {
+            UDPServer us = (UDPServer) server;
+            if(us.needsWrite()) {
+              server.getSelectableChannel().register(selector, SelectionKey.OP_READ|SelectionKey.OP_WRITE);  
             } else {
-              final Server server = servers.get(sk.channel());
-              if(server != null) {
-                if(sk.isReadable()) {
-                  final DatagramChannel dgc = (DatagramChannel) sk.channel();
-                  server.acceptChannel(dgc);
-                }
-              }
+              server.getSelectableChannel().register(selector, SelectionKey.OP_READ);
             }
           }
+        } catch(Exception e) {
+          IOUtils.closeQuietly(server);
         }
-        if(isRunning()) {
-          readScheduler.execute(this);
+        server = serversToAdd.poll();  
+      }
+      
+      server = serversToRemove.poll();
+      while(server != null) {
+        SelectionKey sk = server.getSelectableChannel().keyFor(selector);
+        if(sk != null) {
+          sk.cancel();
         }
+        server = serversToRemove.poll();
       }
     }
-  }
-
-  /**
-   * Runnable for the Write thread.  This runs the writeSelector on the WriteThread. 
-   */
-  private class WriteRunner implements Runnable {
-    @Override
-    public void run() {
-      if(isRunning()) {
+    
+    private void processClients() {
+      Client client = clientsToCheck.poll();
+      while(client != null) {
+        final Client fc = client;
         try {
-          writeSelector.selectedKeys().clear();
-          writeSelector.select();
+          SelectionKey sk = client.getChannel().keyFor(selector);
+          if(client.isClosed()) {
+            clients.remove(client.getChannel());
+            if(sk != null) {
+              sk.cancel();
+            }
+            client.getClientsThreadExecutor().execute(()->IOUtils.closeQuietly(fc.getChannel()));
+          } else {
+            if(sk == null) {
+              sk = client.getChannel().register(selector, 0);
+            }
+            if(!client.getChannel().isConnected() && client.getChannel().isConnectionPending()) {
+              sk.interestOps(SelectionKey.OP_CONNECT);
+            } else if(client.canWrite() && client.canRead()) {
+              sk.interestOps(SelectionKey.OP_READ|SelectionKey.OP_WRITE);
+            } else if (client.canRead()){
+              sk.interestOps(SelectionKey.OP_READ);
+            } else if (client.canWrite()){
+              sk.interestOps(SelectionKey.OP_WRITE);
+            } else {
+              sk.interestOps(0);
+            }
+          }
+        } catch (CancelledKeyException e) {
+          addClient(fc);
         } catch (Exception e) {
           ExceptionUtils.handleException(e);
-          stopIfRunning();
-          return;
+          IOUtils.closeQuietly(client);
         }
-        if(isRunning() && ! writeSelector.selectedKeys().isEmpty()) {
-          for(final SelectionKey sk: writeSelector.selectedKeys()) {
-            final Client client = clients.get(sk.channel());
-            if(client != null) {
-              doClientWrite(client, writeSelector);
+        client = clientsToCheck.poll();
+      }
+    }
+
+    private void doSelect() {
+      while(isRunning()) {
+      try {
+        isAwake = false;
+        processClients();
+        processServers();
+        selector.selectedKeys().clear();
+        selector.select();
+        isAwake = true;
+        for(final SelectionKey key: selector.selectedKeys()) {
+          try {
+            if(key.isAcceptable()) {
+              key.interestOps(0);
+              schedulerPool.execute(()->{
+                Server s = servers.get(key.channel());
+                doServerAccept(s);
+                addServer(s);
+              });
             } else {
-              final Server server = servers.get(sk.channel());
-              if(server != null) {
-                if(server instanceof UDPServer) {
-                  UDPServer us = (UDPServer) server;
-                  us.doWrite();
-                  setUDPServerOperations(us, true);
+              final Client tmpClient = clients.get(key.channel());
+              if(key.isConnectable() && tmpClient != null) {
+                doClientConnect(tmpClient, selector);
+                key.cancel(); //Stupid windows bug here.
+              } else {
+                if (key.isReadable()) {
+                  if(tmpClient != null){
+                    doClientRead(tmpClient, selector);
+                  } else {
+                    final Server server = servers.get(key.channel());
+                    if(server != null && server.getServerType() == WireProtocol.UDP) {
+                      server.acceptChannel((DatagramChannel)server.getSelectableChannel());
+                    }
+                  }
+                } 
+                if(key.isWritable()) {
+                  if(tmpClient != null){
+                    doClientWrite(tmpClient, selector);
+                  } else {
+                    final Server server = servers.get(key.channel());
+                    if(server != null) {
+                      if(server instanceof UDPServer) {
+                        UDPServer us = (UDPServer) server;
+                        stats.addWrite(us.doWrite());
+                        setUDPServerOperations(us, true);
+                      }
+                    }
+                  }
                 }
               }
             }
+          } catch(CancelledKeyException e) {
+            //Key could be cancelled at any point, we dont really care about it.
           }
         }
-        if(isRunning()) {
-          writeScheduler.execute(this);
-        }
+        //Also for windows bug, canceled keys are not removed till we select again.
+        //So we just have to at the end of the loop.
+        selector.selectNow();
+      } catch (IOException e) {
+        //There is really nothing to do here but try again, usually this is because of shutdown.
+      } catch(ClosedSelectorException e) {
+        //We do nothing here because the next loop should not happen now.
+      } catch (NullPointerException e) {
+        //There is a bug in some JVMs around this where the select() can throw an NPE from native code.
+      }
       }
     }
-  }
-
-  @Override
-  public SubmitterExecutor getExecutorFor(final Object obj) {
-    return clientDistributer.getExecutorForKey(obj);
   }
 }
